+++ src/main/java/org/example/models/RangeAnswer.java
@@ -17,7 +17,8 @@ public class RangeAnswer {
 
     private int q_id;
 
-    private int submission_id;
+
+    private long submission_id;
 
     private int answer;
 }
